Implement path following for frosh
Each frosh holds a reference to a path index, corresponding
to a position on the path. Each call to update() will move the
frosh towards that point, according to its internal speed. In
addition, logic to decrease the health when a frosh reaches
the final path index is added.
TBR-47